Tip korisnika: 3 i tacka 2.4: Administrator avio kompanije i njegov profil
Izmena avio kompanije, dodavanje i brisanje destinacija, dodavanje, izmena i brisanje letova, izmena podataka administratora

+++ projectISA/src/main/java/ftn/isa/controller/AvioCompanyController.java
@@ -1,9 +1,13 @@
 package ftn.isa.controller;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import javax.mail.search.AddressTerm;
+
+import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import ftn.isa.dto.AddressDTO;
 import ftn.isa.dto.AvioCompanyDTO;
 import ftn.isa.dto.AvioFlightDTO;
 import ftn.isa.model.Address;
@@ -51,12 +56,14 @@ public class AvioCompanyController {
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
 	public ResponseEntity<AvioCompanyDTO> getAvioCompany(@PathVariable("id") Long id) {
-		AvioCompany avioCopmany = avioCompanyService.findAvioCompany(id);
+		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
 
-		if (avioCopmany == null) {
+		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
-		return new ResponseEntity<>(new AvioCompanyDTO(avioCopmany), HttpStatus.OK);
+		return new ResponseEntity<>(new AvioCompanyDTO(avioCompany), HttpStatus.OK);
 	}
 
 	@RequestMapping(method = RequestMethod.POST, consumes = "application/json")
@@ -73,24 +80,26 @@ public class AvioCompanyController {
 		address.setPostalCode(avioCompanyDTO.getAddress().getPostalCode());
 		address.setStreet(avioCompanyDTO.getAddress().getStreet());
 		address.setNumber(avioCompanyDTO.getAddress().getNumber());
-		
+
 		address = addressService.save(address);
-		
+
 		avioCompany.setName(avioCompanyDTO.getName());
 		avioCompany.setDescription(avioCompanyDTO.getDescription());
 		avioCompany.setAddress(address);
 		avioCompany.setAverageRating(avioCompanyDTO.getAverageRating());
-
+		avioCompany.setDeleted(false);
 		avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
 
 		return new ResponseEntity<>(new AvioCompanyDTO(avioCompany), HttpStatus.OK);
 	}
 
-	@RequestMapping(method = RequestMethod.PUT, consumes = "application/json")
+	@RequestMapping(value = "/{id}/update", method = RequestMethod.POST, consumes = "application/json")
 	public ResponseEntity<AvioCompanyDTO> updateAvioCompnay(@RequestBody AvioCompanyDTO avioCompanyDTO) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(avioCompanyDTO.getId());
 		if (avioCompany == null) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 		avioCompany.setName(avioCompanyDTO.getName());
 		avioCompany.setDescription(avioCompanyDTO.getDescription());
@@ -104,7 +113,7 @@ public class AvioCompanyController {
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Void> removeAvioCompany(@PathVariable("id") Long id) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
-		if (avioCompany != null) {
+		if (avioCompany != null && avioCompany.isDeleted() == false) {
 			avioCompanyService.removeAvioComapny(avioCompany.getId());
 			return new ResponseEntity<>(HttpStatus.OK);
 		} else {
@@ -114,44 +123,42 @@ public class AvioCompanyController {
 
 	// kontroler avio letova
 
-	
 	@RequestMapping(value = "/{id}/flights", method = RequestMethod.GET)
 	public ResponseEntity<List<AvioFlightDTO>> getFlightsOfAviocompany(@PathVariable("id") Long id) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
 
 		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
-		
+
 		Set<Address> destinations = avioCompany.getDestinations();
 		List<AvioFlight> flights = avioCompany.getFlights();
 
-		if (flights.isEmpty()) {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
-
 		List<AvioFlightDTO> flightsDTO = new ArrayList<AvioFlightDTO>();
 
-		if(destinations.isEmpty()) {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);			
-		}else {
+		if (!destinations.isEmpty() && !flights.isEmpty()) {
 			for (AvioFlight avioFlight : flights) {
-				//provera da li destinacije te avio kompanije sadrze polaznu i krajnju tacku leta
-				if(destinations.contains(avioFlight.getStartLocation()) && destinations.contains(avioFlight.getEndLocation())) {
-					flightsDTO.add(new AvioFlightDTO(avioFlight));		
+				if (destinations.contains(avioFlight.getStartLocation())
+						&& destinations.contains(avioFlight.getEndLocation())) {
+					if (!avioFlight.isDeleted()) {
+						flightsDTO.add(new AvioFlightDTO(avioFlight));
+					}
 				}
 			}
-			return new ResponseEntity<>(flightsDTO, HttpStatus.OK);
 		}
+		return new ResponseEntity<>(flightsDTO, HttpStatus.OK);
 	}
 
-	
 	@RequestMapping(value = "/{id}/flights/{flightId}", method = RequestMethod.GET)
 	public ResponseEntity<AvioFlightDTO> getOneFlightOfAvioCompany(@PathVariable("id") Long id,
 			@PathVariable("flightId") Long flightId) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
 
 		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 		List<AvioFlight> flights = avioCompany.getFlights();
@@ -162,6 +169,8 @@ public class AvioCompanyController {
 
 		AvioFlight tempFlight = avioFlightService.findAvioFlight(flightId);
 		if (tempFlight == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (tempFlight.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 
@@ -174,7 +183,6 @@ public class AvioCompanyController {
 
 	}
 
-	
 	@RequestMapping(value = "/{id}/flights", method = RequestMethod.POST, consumes = "application/json")
 	public ResponseEntity<AvioFlightDTO> addFlightToAvioCompany(@PathVariable("id") Long id,
 			@RequestBody AvioFlightDTO avioFlightDTO) {
@@ -183,55 +191,66 @@ public class AvioCompanyController {
 		if (avioCompany == null) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
-		List<AvioFlight> flights = avioCompany.getFlights();
+		List<AvioFlight> flights = new ArrayList<AvioFlight>();
 
 		AvioFlight avioFlight = new AvioFlight();
-		avioFlight.setDateTimeStart(avioFlightDTO.getDateTimeStart());
-		avioFlight.setDateTimeFinish(avioFlightDTO.getDateTimeFinish());
-		avioFlight.setFlightDuration(avioFlightDTO.getFlightDuration());
-		avioFlight.setFlightDistance(avioFlightDTO.getFlightDistance());
-		avioFlight.setPrice(avioFlightDTO.getPrice());
+		if (avioCompany.getFlights() != null) {
+			flights = avioCompany.getFlights();
+			Address startLocation = addressService.findOne(avioFlightDTO.getStartLocation().getId());
+			Address endLocation = addressService.findOne(avioFlightDTO.getEndLocation().getId());
+			avioFlight.setStartLocation(startLocation);
+			avioFlight.setEndLocation(endLocation);
+			avioFlight.setDateTimeStart(avioFlightDTO.getDateTimeStart());
+			avioFlight.setDateTimeFinish(avioFlightDTO.getDateTimeFinish());
+			avioFlight.setFlightDuration(avioFlightDTO.getFlightDuration());
+			avioFlight.setFlightDistance(avioFlightDTO.getFlightDistance());
+			avioFlight.setPrice(avioFlightDTO.getPrice());
+			avioFlight.setDeleted(false);
+			avioFlight.setAvioCompany(avioCompany);
+			avioFlight = avioFlightService.saveAvioFlight(avioFlight);
 
-		avioFlight = avioFlightService.saveAvioFlight(avioFlight);
+			flights.add(avioFlight);
+			avioCompany.setFlights(flights);
+			avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
 
-		flights.add(avioFlight);
-		avioCompany.setFlights(flights);
-		avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
+		}
 
 		return new ResponseEntity<>(new AvioFlightDTO(avioFlight), HttpStatus.CREATED);
 
 	}
 
-	
-	@RequestMapping(value = "/{id}/flights/{flightId}", method = RequestMethod.PUT, consumes = "application/json")
+	@RequestMapping(value = "/{id}/flights/{flightId}", method = RequestMethod.POST, consumes = "application/json")
 	public ResponseEntity<AvioFlightDTO> updateFlightOfAvioCompany(@PathVariable("id") Long id,
 			@PathVariable("flightId") Long flightId, @RequestBody AvioFlightDTO avioFlightDTO) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
 
 		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 		List<AvioFlight> flights = avioCompany.getFlights();
 		AvioFlight tempFlight = avioFlightService.findAvioFlight(flightId);
 
 		if (tempFlight == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (tempFlight.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 
-		if (flights.contains(tempFlight)) {
-			flights.remove(tempFlight);
+		if (flights.contains(tempFlight) && !tempFlight.isDeleted()) {
+			Address startLocation = addressService.findOne(avioFlightDTO.getStartLocation().getId());
+			Address endLocation = addressService.findOne(avioFlightDTO.getEndLocation().getId());
 
+			tempFlight.setStartLocation(startLocation);
+			tempFlight.setEndLocation(endLocation);
 			tempFlight.setDateTimeStart(avioFlightDTO.getDateTimeStart());
 			tempFlight.setDateTimeFinish(avioFlightDTO.getDateTimeFinish());
 			tempFlight.setFlightDuration(avioFlightDTO.getFlightDuration());
 			tempFlight.setFlightDistance(avioFlightDTO.getFlightDistance());
 			tempFlight.setPrice(avioFlightDTO.getPrice());
-
-			tempFlight = avioFlightService.saveAvioFlight(tempFlight);
-
-			flights.add(tempFlight);
-			avioCompany.setFlights(flights);
-			avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
+			tempFlight.setDeleted(avioFlightDTO.isDeleted());
+			tempFlight = avioFlightService.updateAvioFlight(tempFlight);
 
 			return new ResponseEntity<>(new AvioFlightDTO(tempFlight), HttpStatus.OK);
 		} else {
@@ -239,19 +258,21 @@ public class AvioCompanyController {
 		}
 	}
 
-	
 	@RequestMapping(value = "/{id}/flights/{flightId}", method = RequestMethod.DELETE)
 	public ResponseEntity<Void> removeFlightOfAvioCompany(@PathVariable("id") Long id,
 			@PathVariable("flightId") Long flightId) {
 		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
 
 		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		} else if (avioCompany.isDeleted() == true) {
+			System.out.println("OVDE PUKO");
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 		List<AvioFlight> flights = avioCompany.getFlights();
 		AvioFlight tempFlight = avioFlightService.findAvioFlight(flightId);
 
-		if (tempFlight == null) {
+		if (tempFlight == null || tempFlight.isDeleted() == true) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		} else {
 			if (flights.contains(tempFlight)) {
@@ -263,6 +284,84 @@ public class AvioCompanyController {
 			} else {
 				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 			}
+		}
+	}
+
+	@RequestMapping(value = "/{id}/destinations", method = RequestMethod.GET)
+	public ResponseEntity<List<AddressDTO>> getDestinationsOfAvioCompany(@PathVariable("id") Long id) {
+		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
+		Set<Address> destinations = avioCompany.getDestinations();
+
+		List<AddressDTO> destinationsDTO = new ArrayList<AddressDTO>();
+
+		if (!destinations.isEmpty()) {
+			for (Address destinatnion : destinations) {
+				destinationsDTO.add(new AddressDTO(destinatnion));
+			}
+		}
+		return new ResponseEntity<>(destinationsDTO, HttpStatus.OK);
+
+	}
+
+	@RequestMapping(value = "/{id}/destination", method = RequestMethod.POST, consumes = "application/json")
+	public ResponseEntity<AddressDTO> addDestination(@PathVariable("id") Long id,
+			@RequestBody AddressDTO destinationDTO) {
+		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
+		if (avioCompany == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+
+		Set<Address> destinations = new HashSet<Address>();
+
+		Address tempAddress = new Address();
+
+		if (avioCompany.getDestinations() != null) {
+			destinations = avioCompany.getDestinations();
+			
+			tempAddress.setCountry(destinationDTO.getCountry());
+			tempAddress.setCity(destinationDTO.getCity());
+			tempAddress.setPostalCode(destinationDTO.getPostalCode());
+			tempAddress.setStreet(destinationDTO.getStreet());
+			tempAddress.setNumber(destinationDTO.getNumber());
+			
+			tempAddress = addressService.save(tempAddress);
+			destinations.add(tempAddress);
+			avioCompany.setDestinations(destinations);
+			avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
+			
+		}
+
+		return new ResponseEntity<>(new AddressDTO(tempAddress), HttpStatus.OK);
+	}
+
+	@RequestMapping(value = "/{id}/removeDestination/{idDestination}", method = RequestMethod.DELETE)
+	public ResponseEntity<Void> removeDestinationOfAvioCompany(@PathVariable("id") Long id,
+			@PathVariable("idDestination") Long idDestination) {
+		AvioCompany avioCompany = avioCompanyService.findAvioCompany(id);
+		Set<Address> destinations = avioCompany.getDestinations();
+		if (destinations.isEmpty()) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+
+		Address destination = addressService.findOne(idDestination);
+		if (destinations.contains(destination)) {
+			List<AvioFlight> flights = avioCompany.getFlights();
+			if (!flights.isEmpty()) {
+				List<AvioFlight> tempFlights = avioCompany.getFlights();
+				for (AvioFlight flight : tempFlights) {
+					if ((flight.getStartLocation().getId().equals(destination.getId()))
+							|| flight.getEndLocation().getId().equals(destination.getId())) {
+						System.out.println("USO OVDE");
+						avioFlightService.removeAvioFlight(flight.getId());
+						// avioCompany.getFlights().remove(flight);
+					}
+				}
+			}
+			avioCompany.getDestinations().remove(destination);
+			avioCompany = avioCompanyService.saveAvioCompany(avioCompany);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
 	}
